cloud_functions/big_query_ddl/requirements.txt
functions-framework==3.*
google-cloud-bigquery
--extra-index-url https://us-python.pkg.dev/du-labs/test-repo/simple
gcp-common-utils